Isolated database representation code
